Change double to float for boundary pos and energy

+++ include/WCSimRootEvent.hh
@@ -43,8 +43,8 @@ private:
   Double_t fTime;
   Int_t fId;
   Int_t fParentId;
-  std::vector<std::vector<double>> boundaryPoints;
-  std::vector<double> boundaryKEs;
+  std::vector<std::vector<float>> boundaryPoints;
+  std::vector<float> boundaryKEs;
   std::vector<int> boundaryTypes; // 1 = blacksheet, 2 = tyvek, 3 = cave
 
 public:
@@ -64,8 +64,8 @@ public:
 		 Double_t time,
 		 Int_t id,
 		 Int_t idParent,
-     std::vector<std::vector<double>> bPs,
-     std::vector<double> bKEs,
+     std::vector<std::vector<float>> bPs,
+     std::vector<float> bKEs,
      std::vector<int> bTypes);
   virtual ~WCSimRootTrack() { }
   bool CompareAllVariables(const WCSimRootTrack * c) const;
@@ -85,8 +85,8 @@ public:
   Double_t   GetTime() const { return fTime;}
   Int_t     GetId() const {return fId;}
   Int_t     GetParentId() const {return fParentId;}
-  std::vector<std::vector<double>> GetBoundaryPoints() {return boundaryPoints;}
-  std::vector<double> GetBoundaryKEs() {return boundaryKEs;}
+  std::vector<std::vector<float>> GetBoundaryPoints() {return boundaryPoints;}
+  std::vector<float> GetBoundaryKEs() {return boundaryKEs;}
   std::vector<int> GetBoundaryTypes() {return boundaryTypes;}
 
   ClassDef(WCSimRootTrack,4)
@@ -457,8 +457,8 @@ public:
 				   Double_t time,
 				   Int_t id,
 				   Int_t idParent,
-           std::vector<std::vector<double>> bPs,
-           std::vector<double> bKEs,
+           std::vector<std::vector<float>> bPs,
+           std::vector<float> bKEs,
            std::vector<int> bTypes);
 
   WCSimRootTrack * AddTrack   (WCSimRootTrack * track);
